refactoring: moved web sockets service and context into appropriate sub-folder (hooks)
