+++ packages/client/src/index.tsx
@@ -1,7 +1,7 @@
 import React, { Suspense } from 'react';
 import { render } from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
-import { SocketProvider } from "~/components/App/socket-context";
+import { SocketProvider } from '~/components/hooks/web-socket/socket-context';
 
 import { APP_ROOT } from '~/config';
 import { App } from '~/components/app';
@@ -14,7 +14,7 @@ function ReactApp(): JSX.Element {
   return (
     <Suspense fallback={<div />}>
       <SocketProvider>
-      <App />
+        <App />
       </SocketProvider>
     </Suspense>
   );
